just in case

+++ GetCompleteGraphWithMinimumDistanceBetweenPairs.m
@@ -3,7 +3,7 @@ function [ adjacencyMatrix ] = GetCompleteGraphWithMinimumDistanceBetweenPairs( 
 %   Detailed explanation goes here
 
     mDistanceBetweenClusters = squareform(distanceBetweenClusters);
-    mDistanceBetweenClusters(logical(eye(size(mDistanceBetweenClusters)))) = intmax('int16');
+    mDistanceBetweenClusters(logical(eye(size(mDistanceBetweenClusters)))) = intmax('int32');
 
 	for i=1:size(adjacencyMatrix, 1)
         minimumDistance = min(mDistanceBetweenClusters(i,:));
@@ -19,8 +19,8 @@ function [ adjacencyMatrix ] = GetCompleteGraphWithMinimumDistanceBetweenPairs( 
         end
         
         
-        mDistanceBetweenClusters(rowMin, colMin) = intmax('int16');
-        mDistanceBetweenClusters(colMin, rowMin) = intmax('int16');
+        mDistanceBetweenClusters(rowMin, colMin) = intmax('int32');
+        mDistanceBetweenClusters(colMin, rowMin) = intmax('int32');
 		
 		if graphconncomp(adjacencyMatrix, 'Directed', 'false') == 1
 			break
