+++ GetCompleteGraphWithMinimumDistances.m
@@ -3,7 +3,7 @@ function [ adjacencyMatrix ] = GetCompleteGraphWithMinimumDistances( distanceBet
 %   Detailed explanation goes here
 
     mDistanceBetweenClusters = squareform(distanceBetweenClusters);
-    mDistanceBetweenClusters(logical(eye(size(mDistanceBetweenClusters)))) = intmax('int16');
+    mDistanceBetweenClusters(logical(eye(size(mDistanceBetweenClusters)))) = intmax('int32');
 
     while graphconncomp(adjacencyMatrix, 'Directed', 'false') > 1
         minimumDistance = min(mDistanceBetweenClusters(:));
@@ -19,8 +19,8 @@ function [ adjacencyMatrix ] = GetCompleteGraphWithMinimumDistances( distanceBet
         end
         
         
-        mDistanceBetweenClusters(rowMin, colMin) = intmax('int16');
-        mDistanceBetweenClusters(colMin, rowMin) = intmax('int16');
+        mDistanceBetweenClusters(rowMin, colMin) = intmax('int32');
+        mDistanceBetweenClusters(colMin, rowMin) = intmax('int32');
     end
 end
